refactor: simplify API fetch error handling in aa-chains route
