+++ apps/portal/src/components/Document/AAChainList.tsx
@@ -1,19 +1,45 @@
 /* eslint-disable @next/next/no-img-element */
 import { cn } from "@/lib/utils";
 import type { ChainMetadata } from "thirdweb/chains";
-import { getBaseUrl } from "../../lib/getBaseUrl";
+
+type ApiResponseType = {
+  data: Record<string, { service: string; enabled: boolean }[]>;
+};
 
 async function getChains(): Promise<ChainMetadata[]> {
   try {
-    const chains = await fetch(`${getBaseUrl()}/api/aa-chains`);
-    if (!chains.ok) {
-      return [];
-    }
-    const result = (await chains.json()) as { data: ChainMetadata[] };
-    return result.data;
+    const [chainsWithServices, allChains] = await Promise.all([
+      fetch("https://api.thirdweb.com/v1/chains/services", {
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }).then((res) => res.json() as Promise<ApiResponseType>),
+      fetch("https://api.thirdweb.com/v1/chains", {
+        headers: {
+          "Content-Type": "application/json",
+        },
+      }).then((res) => res.json() as Promise<{ data: ChainMetadata[] }>),
+    ]);
+
+    const aaChains = Object.entries(chainsWithServices.data)
+      .filter(([, services]) =>
+        services.some(
+          (service) =>
+            service.service === "account-abstraction" && service.enabled,
+        ),
+      )
+      .map(([chainId]) => Number(chainId));
+
+    const intersectedChains = allChains.data
+      .filter((chain) =>
+        aaChains.some((aaChainId) => aaChainId === chain.chainId),
+      )
+      .filter((c) => c.name)
+      .sort((a, b) => a.name.localeCompare(b.name));
+    return intersectedChains;
   } catch (error) {
-    console.error(error);
-    return [];
+    console.error("Failed to fetch chains", error);
+    throw error;
   }
 }
